backend/requirements.txt
Django>=1.10.6
psycopg2
djangorestframework>=3.6.1
django-cors-headers